# artifacts/openapi/wecms-api-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WeCMS API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/system/roles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "GET /api/v1/system/roles",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PagedRoleSummary"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "access-control",
        "x-wecms-permission": "sys:role:list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "POST /api/v1/system/roles",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RoleMutationResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "access-control",
        "x-wecms-permission": "sys:role:create",
        "x-wecms-audit": {
          "module": "access-control",
          "resource": "roles",
          "action": "create"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
    "/api/v1/system/roles/{id:long}": {
      "delete": {
        "tags": [
          "Roles"
        ],
        "summary": "DELETE /api/v1/system/roles/{id:long}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "access-control",
        "x-wecms-permission": "sys:role:delete",
        "x-wecms-audit": {
          "module": "access-control",
          "resource": "roles",
          "action": "delete"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      },
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "GET /api/v1/system/roles/{id:long}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RoleDetailDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "access-control",
        "x-wecms-permission": "sys:role:detail"
      },
      "put": {
        "tags": [
          "Roles"
        ],
        "summary": "PUT /api/v1/system/roles/{id:long}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RoleMutationResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "access-control",
        "x-wecms-permission": "sys:role:update",
        "x-wecms-audit": {
          "module": "access-control",
          "resource": "roles",
          "action": "update"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
    "/api/v1/system/roles/{id:long}/disable": {
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "POST /api/v1/system/roles/{id:long}/disable",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "access-control",
        "x-wecms-permission": "sys:role:disable",
        "x-wecms-audit": {
          "module": "access-control",
          "resource": "roles",
          "action": "disable"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
    "/api/v1/system/roles/{id:long}/enable": {
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "POST /api/v1/system/roles/{id:long}/enable",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "access-control",
        "x-wecms-permission": "sys:role:enable",
        "x-wecms-audit": {
          "module": "access-control",
          "resource": "roles",
          "action": "enable"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
    "/api/v1/system/roles/{id:long}/menus": {
      "put": {
        "tags": [
          "Roles"
        ],
        "summary": "PUT /api/v1/system/roles/{id:long}/menus",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleMenusRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "access-control",
        "x-wecms-permission": "sys:role:assign-menu",
        "x-wecms-audit": {
          "module": "access-control",
          "resource": "roles",
          "action": "assign-menu"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
    "/api/v1/system/roles/{id:long}/permissions": {
      "put": {
        "tags": [
          "Roles"
        ],
        "summary": "PUT /api/v1/system/roles/{id:long}/permissions",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRolePermissionsRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "access-control",
        "x-wecms-permission": "sys:role:assign-permission",
        "x-wecms-audit": {
          "module": "access-control",
          "resource": "roles",
          "action": "assign-permission"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
...
    "/api/v1/system/users/{id:long}/roles": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "PUT /api/v1/system/users/{id:long}/roles",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignUserRolesRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "identity",
        "x-wecms-permission": "sys:user:assign-role",
        "x-wecms-audit": {
          "module": "identity",
          "resource": "users",
          "action": "assign-role"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResult": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "nullable": true
          },
          "traceId": {
            "type": [
              "string",
              "null"
            ]
          },
          "fieldErrors": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "Object": {
        "type": "object",
        "additionalProperties": true
      },
...
      "AssignUserRolesRequest": {
        "type": "object",
        "required": [
          "roleIds"
        ],
        "properties": {
          "roleIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "PagedRoleSummary": {
        "type": "object",
        "required": [
          "records",
          "page",
          "pageSize",
          "total"
        ],
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleSummaryDto"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "RoleSummaryDto": {
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "status",
          "isBuiltin",
          "isLocked",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "isBuiltin": {
            "type": "boolean"
          },
          "isLocked": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RoleDetailDto": {
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "status",
          "isBuiltin",
          "isLocked",
          "permissionIds",
          "menuIds",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "isBuiltin": {
            "type": "boolean"
          },
          "isLocked": {
            "type": "boolean"
          },
          "permissionIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "menuIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateRoleRequest": {
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "permissionIds": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "menuIds": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "UpdateRoleRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "AssignRolePermissionsRequest": {
        "type": "object",
        "required": [
          "permissionIds"
        ],
        "properties": {
          "permissionIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "AssignRoleMenusRequest": {
        "type": "object",
        "required": [
          "menuIds"
        ],
        "properties": {
          "menuIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "RoleMutationResponse": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}